Make everything use the ssh() routine (and add but don't use the run() routine)

+++ silverlining/commands/setup_node.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 from cmdutils import CommandError
+from silverlining.util import ssh
 
 def setup_rsync(config, source, dest):
     cwd = os.path.abspath(os.path.join(__file__, '../../server-files'))
@@ -25,17 +26,7 @@ def command_setup_node(config):
     node = config.args.node
     config.logger.notify(
         'Setting up authentication on server...')
-    ssh_host = 'root@%s' % node
-    proc = subprocess.Popen([
-        'ssh', '-o', 'StrictHostKeyChecking=no',
-        ssh_host, '''
-if [ -e /root/.silverlining-server-setup ] ; then
-    exit 50
-fi
-mkdir -p /root/.ssh
-cat >> /root/.ssh/authorized_keys
-''',
-        ], stdin=subprocess.PIPE)
+    
     for path in ['id_rsa.pub', 'id_dsa.pub']:
         pubkey_path = os.path.join(os.environ['HOME'], '.ssh', path)
         if os.path.exists(pubkey_path):
@@ -46,25 +37,27 @@ cat >> /root/.ssh/authorized_keys
         config.logger.fatal("Can't locate any key file")
         #not sure what error code are used for here but 8 was unused
         return 8
-    proc.communicate(key)
-    # if proc.returncode == 50:
-    #     config.logger.fatal(
-    #         "The server has already been setup (/root/.silverlining-server-setup exists)")
-    #     return 2
+    ssh('root', node, '''
+if [ -e /root/.silverlining-server-setup ] ; then
+    exit 50
+fi
+mkdir -p /root/.ssh
+cat >> /root/.ssh/authorized_keys
+''',
+        ssh_args=['-o', 'StrictHostKeyChecking=no'],
+        stdin=key)
     config.logger.notify(
         "Updating indexes and setting up rsync")
-    proc = subprocess.Popen([
-        'ssh', ssh_host, '''
+    stdout, stderr, returncode = ssh(
+        'root', node, '''
 dpkg --configure -a
 apt-get update -qq
 apt-get -y -q install rsync
-''',
-        ])
-    proc.communicate()
-    if proc.returncode:
+''')
+    if returncode:
         config.logger.fatal(
             "An error occurred (code=%r)"
-            % proc.returncode)
+            % returncode)
         response = config.ask(
             "Continue?")
         if not response:
@@ -76,25 +69,20 @@ apt-get -y -q install rsync
     packages = ' '.join(line.strip().split()[0]
                         for line in lines
                         if line.strip())
-    proc = subprocess.Popen([
-        'ssh', ssh_host,
-        'apt-get -y -q=2 install $(cat)'],
-                            stdin=subprocess.PIPE)
-    proc.communicate(packages)
-    if proc.returncode:
+    stdout, stderr, returncode = ssh(
+        'root', node, 'apt-get -y -q=2 install $(cat)',
+        stdin=packages)
+    if returncode:
         config.logger.fatal(
             "An error occurred (code=%r)"
-            % proc.returncode)
+            % returncode)
         response = config.ask(
             "Continue?")
         if not response:
             return 5
     
     setup_rsync(config, 'serverroot/', '/')
-    proc = subprocess.Popen([
-        'ssh', ssh_host,
-        'mv /var/root/* /root/'])
-    proc.communicate()
+    ssh('root', node, 'mv /var/root/* /root/')
     
     # Move over the root files, we do *not* rsync a /root dir because
     # that would copy the wrong permissions for the /root directorty
@@ -106,9 +94,9 @@ apt-get -y -q install rsync
     import getpass
     username = getpass.getuser()
     setup_script = setup_script.replace('__REMOTE_USER__', username)
-    
-    stdout, returncode = config.run(
-        ['ssh', ssh_host, setup_script])
+
+    stdout, stderr, returncode = ssh(
+        'root', node, setup_script)
     if returncode:
         config.logger.fatal(
             "An error occurred (code=%r)"
